// frontend/src/app/api/model/createModuleElementDto.ts
/**
 * CoreData API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ModuleElementType } from './moduleElementType';


export interface CreateModuleElementDto { 
    type?: ModuleElementType;
    sortOrder?: number;
    iconKey?: string | null;
    options?: any | null;
    assessmentWeight?: number | null;
}
export namespace CreateModuleElementDto {
}