// frontend/src/app/api/model/moduleElementDto.ts
/**
 * CoreData API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ModuleElementType } from './moduleElementType';


export interface ModuleElementDto { 
    id?: string;
    sortOrder?: number;
    type?: ModuleElementType;
    iconKey?: string | null;
    options?: any | null;
    assessmentWeight?: number | null;
    marksPublished?: boolean;
}
export namespace ModuleElementDto {
}