// frontend/src/app/api/model/moduleFileDto.ts
/**
 * CoreData API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ModuleFileDto { 
    id?: string;
    moduleId?: string;
    originalFileName?: string | null;
    contentType?: string | null;
    sizeBytes?: number;
}