// frontend/src/app/api/model/updateModuleElementDto.ts
/**
 * CoreData API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UpdateModuleElementDto { 
    sortOrder?: number;
    iconKey?: string | null;
    options?: any | null;
    assessmentWeight?: number | null;
}